+++ cooked-validators/src/Cooked.hs
@@ -10,4 +10,5 @@ import Cooked.Output as X
 import Cooked.Pretty as X
 import Cooked.RawUPLC as X
 import Cooked.Skeleton as X
+import Cooked.Tweak as X
 import Cooked.Wallet as X
