+++ cooked-validators/src/Cooked/Attack.hs
@@ -4,4 +4,3 @@ import Cooked.Attack.AddToken as X
 import Cooked.Attack.DatumHijacking as X
 import Cooked.Attack.DoubleSat as X
 import Cooked.Attack.DupToken as X
-import Cooked.Attack.Tweak as X
