+++ cooked-validators/src/Cooked/Attack/DatumHijacking.hs
@@ -11,10 +11,10 @@
 module Cooked.Attack.DatumHijacking where
 
 import Control.Monad
-import Cooked.Attack.Tweak
-import Cooked.RawUPLC
 import Cooked.Output
+import Cooked.RawUPLC
 import Cooked.Skeleton
+import Cooked.Tweak
 import qualified Ledger as L
 import qualified Ledger.Typed.Scripts as L
 import Optics.Core
