+++ cooked-validators/src/Cooked/Attack/DoubleSat.hs
@@ -7,10 +7,10 @@
 module Cooked.Attack.DoubleSat where
 
 import Control.Monad
-import Cooked.Attack.Tweak
 import Cooked.MockChain.Monad
 import Cooked.Output
 import Cooked.Skeleton
+import Cooked.Tweak
 import Cooked.Wallet
 import Data.List
 import Data.Map (Map)
