--- cooked-validators/src/Cooked/Attack/Tweak.hs
@@ -1,6 +1,0 @@
-module Cooked.Attack.Tweak (module X) where
-
-import Cooked.Attack.Tweak.AddInputsAndOutputs as X
-import Cooked.Attack.Tweak.Common as X
-import Cooked.Attack.Tweak.OutPermutations as X
-import Cooked.Attack.Tweak.TamperDatum as X
